cleaned up the code for launcher.py
