+++ tempCodeRunnerFile.py
@@ -1,1 +1,2 @@
-from logtool import Logger
+struct_commands) 
+        self.button_box.rejected.connect(self.close)
